pyproject.toml
[project]
name = "crm"
version = "0.1.0"
description = ""
authors = [
    {name = "Zondar__",email = "kirito98730tkm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
package-mode = false

dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
